Update MFA_NPS_Troubleshooter.ps1
The script should not make changes to the registry prior to the prompt that describes what it is doing. If the user presses Ctrl+C to cancel the script, changes would have already been made to the registry which are not obvious which contradicts the "Press ENTER to continue"... message.